+++ src/api/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.16.0] - 2024-12-12
+
 ### Changed
 
 - Upgrade fastapi to `0.115.6`
@@ -256,7 +258,8 @@ and this project adheres to
 
 - Implement base FastAPI app
 
-[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.15.0...main
+[unreleased]: https://github.com/MTES-MCT/qualicharge/compare/v0.16.0...main
+[0.16.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/MTES-MCT/qualicharge/compare/v0.12.1...v0.13.0